pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mavrith"
version = "0.1.0"
description = "A lightweight block-based simulation framework for Python."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" },
]
dependencies = [
  "scipy>=1.13",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]
yaml = [
  "PyYAML>=6",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src", "."]
testpaths = ["tests"]
markers = [
  "performance: performance-oriented tests for large generated models",
  "slow: slower-running tests for large generated models",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mavrith = ["py.typed"]
